+++ TaskweaveWidget/TaskweaveWidgetBundle.swift
@@ -1,0 +1,9 @@
+import WidgetKit
+import SwiftUI
+
+@main
+struct TaskweaveWidgetBundle: WidgetBundle {
+    var body: some Widget {
+        TaskweaveWidget()
+    }
+}